Fixed the play/pause error in logs
Bugs still present.

1. Since the bus stop/journey/return sounds
are pretty similar, it is hard to tell what
we are listening to at any given point. If
there was a message, say on the bottom right
corner of the screen that said what we are
listening to right now, that'll help a lot.
2. The animation events aren't synchronized
with the audio events. For example, say I
click on bus stop A. I'll start hearing the
bus stop sounds. Then I start the journey.
Now the bus moves and I hear the journey sounds.
Now if I again click on A, I hear the bus stop
sounds but the bus keeps moving on the path.
In fact, it is possible to do a sequence of
interactions so that the path disappears
and the bus still moves.